Move errors.ts to /common
