+++ extensions/ql-vscode/src/common/errors.ts
