+++ extensions/ql-vscode/src/common/telemetry.ts
@@ -1,4 +1,4 @@
-import { RedactableError } from "../pure/errors";
+import { RedactableError } from "./errors";
 
 export interface AppTelemetry {
   sendCommandUsage(name: string, executionTime: number, error?: Error): void;
