+++ extensions/ql-vscode/src/common/vscode/telemetry.ts
@@ -18,7 +18,7 @@ import * as appInsights from "applicationinsights";
 import { extLogger } from "../index";
 import { UserCancellationException } from "./progress";
 import { showBinaryChoiceWithUrlDialog } from "./dialog";
-import { RedactableError } from "../../pure/errors";
+import { RedactableError } from "../errors";
 import { SemVer } from "semver";
 import { AppTelemetry } from "../telemetry";
