+++ extensions/ql-vscode/src/data-extensions-editor/external-api-usage-query.ts
@@ -13,7 +13,7 @@ import { ProgressCallback } from "../common/vscode/progress";
 import { fetchExternalApiQueries } from "./queries";
 import { QueryResultType } from "../pure/new-messages";
 import { join } from "path";
-import { redactableError } from "../pure/errors";
+import { redactableError } from "../common/errors";
 import { telemetryListener } from "../common/vscode/telemetry";
 
 export type RunQueryOptions = {
