+++ extensions/ql-vscode/src/language-support/ast-viewer/ast-viewer.ts
@@ -25,7 +25,7 @@ import {
 } from "../../common/bqrs-utils";
 import { DisposableObject } from "../../common/disposable-object";
 import { asError, getErrorMessage } from "../../pure/helpers-pure";
-import { redactableError } from "../../pure/errors";
+import { redactableError } from "../../common/errors";
 import { AstViewerCommands } from "../../common/commands";
 import { extLogger } from "../../common";
 import { showAndLogExceptionWithTelemetry } from "../../common/logging";
