+++ extensions/ql-vscode/src/language-support/contextual/query-resolver.ts
@@ -22,7 +22,7 @@ import { showAndLogExceptionWithTelemetry } from "../../common/logging";
 import { CancellationToken } from "vscode";
 import { ProgressCallback } from "../../common/vscode/progress";
 import { CoreCompletedQuery, QueryRunner } from "../../query-server";
-import { redactableError } from "../../pure/errors";
+import { redactableError } from "../../common/errors";
 import { QLPACK_FILENAMES } from "../../pure/ql";
 import { telemetryListener } from "../../common/vscode/telemetry";
