+++ extensions/ql-vscode/src/local-queries/local-query-run.ts
@@ -18,7 +18,7 @@ import { CompletedLocalQueryInfo, LocalQueryInfo } from "../query-results";
 import { WebviewReveal } from "./webview";
 import { CodeQLCliServer } from "../codeql-cli/cli";
 import { QueryResultType } from "../pure/new-messages";
-import { redactableError } from "../pure/errors";
+import { redactableError } from "../common/errors";
 import { LocalQueries } from "./local-queries";
 import { tryGetQueryMetadata } from "../codeql-cli/query-metadata";
 import { telemetryListener } from "../common/vscode/telemetry";
