+++ extensions/ql-vscode/src/local-queries/results-view.ts
@@ -72,7 +72,7 @@ import {
 import { isCanary, PAGE_SIZE } from "../config";
 import { HistoryItemLabelProvider } from "../query-history/history-item-label-provider";
 import { telemetryListener } from "../common/vscode/telemetry";
-import { redactableError } from "../pure/errors";
+import { redactableError } from "../common/errors";
 import { ResultsViewCommands } from "../common/commands";
 
 /**
