+++ extensions/ql-vscode/src/packaging/packaging.ts
@@ -12,7 +12,7 @@ import {
   showAndLogInformationMessage,
 } from "../common/logging";
 import { asError, getErrorStack } from "../pure/helpers-pure";
-import { redactableError } from "../pure/errors";
+import { redactableError } from "../common/errors";
 import { PACKS_BY_QUERY_LANGUAGE } from "../common/query-language";
 import { PackagingCommands } from "../common/commands";
 import { telemetryListener } from "../common/vscode/telemetry";
