+++ extensions/ql-vscode/src/query-server/legacy/run-queries.ts
@@ -23,7 +23,7 @@ import * as qsClient from "./query-server-client";
 import { asError, getErrorMessage } from "../../pure/helpers-pure";
 import { compileDatabaseUpgradeSequence } from "./upgrades";
 import { QueryEvaluationInfo, QueryOutputDir } from "../../run-queries-shared";
-import { redactableError } from "../../pure/errors";
+import { redactableError } from "../../common/errors";
 import { CoreQueryResults, CoreQueryTarget } from "../query-runner";
 import { Position } from "../../pure/messages-shared";
 import { ensureDirSync } from "fs-extra";
